LU-3744 llite: use 64bits flags in ll_lov_setea()
In ll_lov_setea(), setting MDS_OPEN_HAS_OBJS to an int flags will
result in the flags being overflowed.

Reviewed-on: http://review.whamcloud.com/7312
Tested-by: Hudson
Tested-by: Maloo <whamcloud.maloo@gmail.com>

+++ lustre/llite/file.c
@@ -1160,7 +1160,8 @@ static int ll_lov_recreate_fid(struct inode *inode, unsigned long arg)
 }
 
 int ll_lov_setstripe_ea_info(struct inode *inode, struct file *file,
-                             int flags, struct lov_user_md *lum, int lum_size)
+                             __u64  flags, struct lov_user_md *lum,
+			     int lum_size)
 {
         struct lov_stripe_md *lsm;
         struct lookup_intent oit = {.it_op = IT_OPEN, .it_flags = flags};
@@ -1275,7 +1276,7 @@ out:
 static int ll_lov_setea(struct inode *inode, struct file *file,
                             unsigned long arg)
 {
-        int flags = MDS_OPEN_HAS_OBJS | FMODE_WRITE;
+        __u64 flags = MDS_OPEN_HAS_OBJS | FMODE_WRITE;
         struct lov_user_md  *lump;
         int lum_size = sizeof(struct lov_user_md) +
                        sizeof(struct lov_user_ost_data);
@@ -1309,7 +1310,7 @@ static int ll_lov_setstripe(struct inode *inode, struct file *file,
         struct lov_user_md_v3 *lumv3p = (struct lov_user_md_v3 *)arg;
         int lum_size;
         int rc;
-        int flags = FMODE_WRITE;
+        __u64 flags = FMODE_WRITE;
         ENTRY;
 
         /* first try with v1 which is smaller than v3 */
